Add files via upload
The goal here is to demo the use of JavaScript and fetch. The app is a concept of a hamburger restaurant. The user goes to the website select a item or two. The selected item goes into the cart. Inside the cart, the user has the option to increase the quantity of a certain item. Once the user finished looking over the menu, the cart filled with item(s) could be easily submitted by pressing the submit button. The cart filled with the customer order goes to the restaurant's database  then restaurant retrieve the the order and prepare the it.